@@ -2,8 +2,9 @@
 {
     public enum ListenerStates
     {
-        ACTIVE,
-        DELETED,
-        PAUSED
+        ACTIVE, // listener actively received recordings
+        DELETED, // listener is soft deleted awaiting for property to be returned
+        PAUSED, // listener temporarily suspended from receiving recordings
+        RESERVED // wallet number is reserved after listener is deleted
     }
 }
