@@ -3,7 +3,6 @@ using System;
 using FluentAssertions;
 using TNBase.Repository;
 using TNBase.Infrastructure.Helpers;
-using Xunit;
 
 namespace TNBase.App.Test
 {
@@ -146,7 +145,7 @@ namespace TNBase.App.Test
         [Fact]
         public void Stats_TotalStoppedWallets()
         {
-            serviceLayer.GetStoppedListeners().Count.Should().Be(1);
+            serviceLayer.GetPausedListeners().Count.Should().Be(1);
         }
 
         [Fact]
